Created ArrayList
Write a small program that asks for the names and IDs of all employees
in a small company, and
store them in an array of integers and an array of Strings. The company
has 10 employees.
Use a loop to go through both arrays and print the names and IDs of
those employees whose
ID is less than 1000. Use another loop to print the names and IDs of
those employees whose name
starts with “J” or “S”.